+++ packages/expo-audio/ios/AudioModule.swift
@@ -185,6 +185,25 @@ public class AudioModule: Module {
         }
       }
 
+      Function("setActiveForLockScreen") { (player: AudioPlayer, active: Bool, metadata: Metadata?) in
+        player.setActiveForLockScreen(active, metadata: metadata)
+      }
+
+      Function("updateLockScreenMetadata") { (player: AudioPlayer, metadata: Metadata?) in
+        if player.isActiveForLockScreen {
+          player.metadata = metadata
+          MediaController.shared.updateNowPlayingInfo(for: player)
+        }
+      }
+
+      Function("clearLockScreenControls") { (player: AudioPlayer) in
+        if player.isActiveForLockScreen {
+          player.metadata = nil
+          player.isActiveForLockScreen = false
+          MediaController.shared.setActivePlayer(nil)
+        }
+      }
+
       AsyncFunction("seekTo") { (player: AudioPlayer, seconds: Double) in
         await player.ref.currentItem?.seek(
           to: CMTime(
@@ -434,7 +453,7 @@ public class AudioModule: Module {
     try AudioUtils.validateAudioMode(mode: mode)
     let session = AVAudioSession.sharedInstance()
     var category: AVAudioSession.Category = session.category
-
+    
     self.shouldPlayInBackground = mode.shouldPlayInBackground
     self.interruptionMode = mode.interruptionMode
     self.allowsRecording = mode.allowsRecording
@@ -481,7 +500,11 @@ public class AudioModule: Module {
       sessionOptions = categoryOptions
     }
 
-    try session.setCategory(category, options: sessionOptions)
+    if sessionOptions.isEmpty {
+      try session.setCategory(category, mode: .default)
+    } else {
+      try session.setCategory(category, options: sessionOptions)
+    }
     try session.setActive(true, options: [.notifyOthersOnDeactivation])
   }